+++ nodeconductor/structure/serializers.py
@@ -22,7 +22,7 @@ from nodeconductor.core.fields import MappedChoiceField
 from nodeconductor.monitoring.serializers import MonitoringSerializerMixin
 from nodeconductor.quotas import serializers as quotas_serializers
 from nodeconductor.structure import (models, SupportedServices, ServiceBackendError, ServiceBackendNotImplemented,
-                                     executors)
+                                     executors, permissions)
 from nodeconductor.structure.managers import filter_queryset_for_user
 from nodeconductor.structure.models import ServiceProjectLink
 
@@ -1162,13 +1162,14 @@ class BaseServiceProjectLinkSerializer(PermissionFieldFilteringMixin,
         read_only=True)
 
     service_name = serializers.ReadOnlyField(source='service.settings.name')
+    quotas = quotas_serializers.BasicQuotaSerializer(many=True, read_only=True)
 
     class Meta(object):
         model = NotImplemented
         fields = (
             'url',
             'project', 'project_name', 'project_uuid',
-            'service', 'service_uuid', 'service_name',
+            'service', 'service_uuid', 'service_name', 'quotas',
         )
         related_paths = ('project', 'service')
         extra_kwargs = {