+++ nodeconductor/structure/tests/models.py
@@ -13,13 +13,10 @@ class TestService(structure_models.Service):
         return 'test'
 
 
-class TestServiceProjectLink(structure_models.ServiceProjectLink):
+class TestServiceProjectLink(structure_models.CloudServiceProjectLink):
     service = models.ForeignKey(TestService)
 
-    class Quotas(QuotaModelMixin.Quotas):
-        vcpu = QuotaField(default_limit=20, is_backend=True)
-        ram = QuotaField(default_limit=51200, is_backend=True)
-        storage = QuotaField(default_limit=1024000, is_backend=True)
+    class Quotas(structure_models.CloudServiceProjectLink.Quotas):
         instances = QuotaField(default_limit=30, is_backend=True)
         security_group_count = QuotaField(default_limit=100, is_backend=True)
         security_group_rule_count = QuotaField(default_limit=100, is_backend=True)